requirements.txt
pillow
curses-util
